[Finishes #108852644] changed test pairing to test_routes
